updated variant filter dto

+++ product-service/src/main/java/com/whitechoke/productservice/api/dto/variantDto/VariantFilterDto.java
@@ -6,7 +6,7 @@ import lombok.Builder;
 public record VariantFilterDto(
         Long id,
         Integer size,
-        Integer productId,
+        Long productId,
         Integer pageSize,
         Integer pageNumber
 ) {
